Added the crioctl tool (used to be crio-status)
crioctl used to be crio-status but the name crio-status seemed wrong
with the addition of checkpoint and restore subcommands.

With crioctl it is possible to run all previous commands that were
possible with crio-status and additionally it has the subcommands:

 * crioctl checkpoint
 * crioctl restore

These two commands are using almost the same parameters as the
checkpoint/restore functionality in Podman.

In addition to the options checkpoint/restore has in the CRI API,
crioctl uses an CRI-O internal/private API with additional
functionality. Especially to restore containers it supports additional
options which make restoring much easier (like re-mapping mount points).

With crioctl it is also possible to test the checkpoint/restore
functionality in CRI-O's provided in test/checkpoint.bats.

To not break crio-status users the Makefile creates symbolic links from
crioctl to crio-status.

Signed-off-by: Adrian Reber <areber@redhat.com>